# src/swagger.json
{
	"components": {
...
		"schemas": {
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
...
			"ErrorName": {
				"type": "string",
				"enum": [
					"VOLUNTEER_NOT_FOUND",
					"VOLUNTEER_ALREADY_EXISTS",
					"VOLUNTEER_NOT_UPDATED",
					"VOLUNTEER_UNREGISTERED",
					"VOLUNTEER_NOT_DELETED",
					"PASSWORD_WRONG_ERROR",
					"INVALID_DATE_REGISTER",
					"HOURS_ALREADY_REGISTERED",
					"HOURS_NOT_FOUND"
				]
			},
...
			"BookEvaluationEntity": {
				"properties": {
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"observedHistories": {
						"type": "string"
					},
					"readHistories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"relevantPhrases": {
						"type": "string"
					},
					"concept": {
						"type": "string"
					},
					"observations": {
						"type": "string"
					},
					"syntheticAvaliation": {
						"type": "string"
					},
					"grammarAvaliation": {
						"type": "string"
					},
					"societyCriticalAnalysisAvaliation": {
						"type": "string"
					},
					"bookCriticalAnalysisAvaliation": {
						"type": "string"
					},
					"textClarityAvaliation": {
						"type": "string"
					},
					"textReliabilityAvaliation": {
						"type": "string"
					},
					"textAestheticsAvaliation": {
						"type": "string"
					},
					"isAppropriation": {
						"type": "boolean"
					},
					"isParcialPlagiarism": {
						"type": "boolean"
					},
					"evaluatorId": {
						"type": "number",
						"format": "double"
					},
					"classId": {
						"type": "number",
						"format": "double"
					},
					"readerRegistration": {
						"type": "number",
						"format": "double"
					},
					"readerName": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"createdAt",
					"readHistories",
					"concept",
					"observations",
					"syntheticAvaliation",
					"grammarAvaliation",
					"societyCriticalAnalysisAvaliation",
					"bookCriticalAnalysisAvaliation",
					"textClarityAvaliation",
					"textReliabilityAvaliation",
					"textAestheticsAvaliation",
					"isAppropriation",
					"isParcialPlagiarism",
					"evaluatorId",
					"classId",
					"readerRegistration",
					"readerName",
					"id"
				],
				"type": "object"
			},
...
			"BookEvaluationList": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BookEvaluationEntity"
					},
					{
						"properties": {
							"volunteerName": {
								"type": "string",
								"nullable": true
							},
							"expirationDate": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							}
						},
						"required": [
							"volunteerName",
							"expirationDate"
						],
						"type": "object"
					}
				]
			},
			"PaginationResult_BookEvaluationList-Array_": {
				"properties": {
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"pageInfo": {
						"properties": {
							"hasPreviousPage": {
								"type": "boolean"
							},
							"hasNextPage": {
								"type": "boolean"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"hasPreviousPage",
							"hasNextPage",
							"page"
						],
						"type": "object"
					},
					"nodes": {
						"items": {
							"$ref": "#/components/schemas/BookEvaluationList"
						},
						"type": "array"
					}
				},
				"required": [
					"totalCount",
					"pageInfo",
					"nodes"
				],
				"type": "object"
			},
			"Pick_BookEvaluationEntity.Exclude_keyofBookEvaluationEntity.id-or-createdAt__": {
				"properties": {
					"readerName": {
						"type": "string"
					},
					"readerRegistration": {
						"type": "number",
						"format": "double"
					},
					"classId": {
						"type": "number",
						"format": "double"
					},
					"evaluatorId": {
						"type": "number",
						"format": "double"
					},
					"isParcialPlagiarism": {
						"type": "boolean"
					},
					"isAppropriation": {
						"type": "boolean"
					},
					"textAestheticsAvaliation": {
						"type": "string"
					},
					"textReliabilityAvaliation": {
						"type": "string"
					},
					"textClarityAvaliation": {
						"type": "string"
					},
					"bookCriticalAnalysisAvaliation": {
						"type": "string"
					},
					"societyCriticalAnalysisAvaliation": {
						"type": "string"
					},
					"grammarAvaliation": {
						"type": "string"
					},
					"syntheticAvaliation": {
						"type": "string"
					},
					"observations": {
						"type": "string"
					},
					"concept": {
						"type": "string"
					},
					"relevantPhrases": {
						"type": "string"
					},
					"readHistories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"observedHistories": {
						"type": "string"
					}
				},
				"required": [
					"readerName",
					"readerRegistration",
					"classId",
					"evaluatorId",
					"isParcialPlagiarism",
					"isAppropriation",
					"textAestheticsAvaliation",
					"textReliabilityAvaliation",
					"textClarityAvaliation",
					"bookCriticalAnalysisAvaliation",
					"societyCriticalAnalysisAvaliation",
					"grammarAvaliation",
					"syntheticAvaliation",
					"observations",
					"concept",
					"readHistories"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_BookEvaluationEntity.id-or-createdAt_": {
				"$ref": "#/components/schemas/Pick_BookEvaluationEntity.Exclude_keyofBookEvaluationEntity.id-or-createdAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateBookEvaluationEntity": {
				"$ref": "#/components/schemas/Omit_BookEvaluationEntity.id-or-createdAt_"
			},
			"BookEvaluationError": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ErrorName"
							},
							{
								"type": "string",
								"enum": [
									"UNSPECIFIED_ERROR"
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_BookEvaluationEntity.Exclude_keyofBookEvaluationEntity.id-or-classId-or-evaluatorId__": {
				"properties": {
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"readerName": {
						"type": "string"
					},
					"readerRegistration": {
						"type": "number",
						"format": "double"
					},
					"isParcialPlagiarism": {
						"type": "boolean"
					},
					"isAppropriation": {
						"type": "boolean"
					},
					"textAestheticsAvaliation": {
						"type": "string"
					},
					"textReliabilityAvaliation": {
						"type": "string"
					},
					"textClarityAvaliation": {
						"type": "string"
					},
					"bookCriticalAnalysisAvaliation": {
						"type": "string"
					},
					"societyCriticalAnalysisAvaliation": {
						"type": "string"
					},
					"grammarAvaliation": {
						"type": "string"
					},
					"syntheticAvaliation": {
						"type": "string"
					},
					"observations": {
						"type": "string"
					},
					"concept": {
						"type": "string"
					},
					"relevantPhrases": {
						"type": "string"
					},
					"readHistories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"observedHistories": {
						"type": "string"
					}
				},
				"required": [
					"createdAt",
					"readerName",
					"readerRegistration",
					"isParcialPlagiarism",
					"isAppropriation",
					"textAestheticsAvaliation",
					"textReliabilityAvaliation",
					"textClarityAvaliation",
					"bookCriticalAnalysisAvaliation",
					"societyCriticalAnalysisAvaliation",
					"grammarAvaliation",
					"syntheticAvaliation",
					"observations",
					"concept",
					"readHistories"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_BookEvaluationEntity.id-or-classId-or-evaluatorId_": {
				"$ref": "#/components/schemas/Pick_BookEvaluationEntity.Exclude_keyofBookEvaluationEntity.id-or-classId-or-evaluatorId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UpdateBookEvaluationEntity": {
				"$ref": "#/components/schemas/Omit_BookEvaluationEntity.id-or-classId-or-evaluatorId_"
			},
...
			"NotebookError": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ErrorName"
							},
							{
								"type": "string",
								"enum": [
									"UNSPECIFIED_ERROR"
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Pick_PepClassEntity.notebookDirectory_": {
				"properties": {
					"notebookDirectory": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NotebookEntity": {
				"allOf": [
					{
						"properties": {
							"reservationDate": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"evaluatedDate": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"archivesExclusion": {
								"type": "boolean"
							},
							"approved": {
								"type": "boolean"
							},
							"conclusion": {
								"type": "string"
							},
							"a13": {
								"type": "string"
							},
							"a12": {
								"type": "string"
							},
							"a11": {
								"type": "string"
							},
							"a10": {
								"type": "string"
							},
							"a9": {
								"type": "string"
							},
							"a8": {
								"type": "string"
							},
							"a7": {
								"type": "string"
							},
							"a6": {
								"type": "string"
							},
							"a5": {
								"type": "string"
							},
							"a4": {
								"type": "string"
							},
							"a3": {
								"type": "string"
							},
							"a2": {
								"type": "string"
							},
							"a1": {
								"type": "string"
							},
							"relevantContent": {
								"type": "string"
							},
							"subject10": {
								"type": "string"
							},
							"subject9": {
								"type": "string"
							},
							"subject8": {
								"type": "string"
							},
							"subject7": {
								"type": "string"
							},
							"subject6": {
								"type": "string"
							},
							"subject5": {
								"type": "string"
							},
							"subject4": {
								"type": "string"
							},
							"subject3": {
								"type": "string"
							},
							"subject2": {
								"type": "string"
							},
							"subject1": {
								"type": "string"
							},
							"evaluatorEmail": {
								"type": "string"
							},
							"evaluatorName": {
								"type": "string"
							},
							"studentPrisonUnit": {
								"type": "string"
							},
							"studentRegistration": {
								"type": "number",
								"format": "double"
							},
							"studentName": {
								"type": "string",
								"nullable": true
							},
							"idpep": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"idvol": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"idcad": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"archivesExclusion",
							"approved",
							"conclusion",
							"evaluatorName",
							"studentRegistration",
							"studentName",
							"idvol",
							"idcad"
						],
						"type": "object"
					},
					{
						"$ref": "#/components/schemas/Pick_PepClassEntity.notebookDirectory_"
					}
				]
			},
...
			"Omit_%2F%2Feslint-disable-next-lineprettier%2FprettierNotebookEntity.-or-idcad-or-studentName-or-studentRegistration-or-idpep-or-reservationDate-or-evaluatedDate-or-notebookDirectory_": {
				"$ref": "#/components/schemas/Pick_NotebookEntity.Exclude_keyofNotebookEntity.-or-idcad-or-studentName-or-studentRegistration-or-idpep-or-reservationDate-or-evaluatedDate-or-notebookDirectory__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"EvaluateNotebookEntity": {
				"$ref": "#/components/schemas/Omit_%2F%2Feslint-disable-next-lineprettier%2FprettierNotebookEntity.-or-idcad-or-studentName-or-studentRegistration-or-idpep-or-reservationDate-or-evaluatedDate-or-notebookDirectory_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "palavrasdepaz",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/book-evaluations": {
			"get": {
				"operationId": "GetVBookEvaluationList",
				"responses": {
					"200": {
						"description": "Successfully generated the metrics",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginationResult_BookEvaluationList-Array_"
								}
							}
						}
					}
				},
				"description": "List all book evaluations.\n\n(The user must have bookPermissionn, which is checked using JWT)\n Pagination\n Page: ?page=number& (page number)\n Limit: ?limit=number& (data quantity - max=30)\n Filter: ?classes=399,404,405& (filter by class ids - separator \",\" (comma))",
				"tags": [
					"Book evaluation"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBookEvaluations",
				"responses": {
					"201": {
						"description": "Successfully Created Evaluations"
					}
				},
				"description": "Create multiple multiple evaluations\n\n(The volunteer must have bookPermission, which is checked using JWT)",
				"tags": [
					"Book evaluation"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/CreateBookEvaluationEntity"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/book-evaluations/download": {
			"get": {
				"operationId": "GetVBookEvaluationListDownload",
				"responses": {
					"200": {
						"description": "Successfully generated the metrics",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"format": "byte"
								}
							}
						}
					}
				},
				"description": "Download list all book evaluations.\n\n(The user must have bookPermissionn, which is checked using JWT)\n Filter: ?classes=399,404,405& (filter by class ids - separator \",\" (comma))",
				"tags": [
					"Book evaluation"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": []
			}
		},
		"/book-evaluations/{evaluationId}": {
			"put": {
				"operationId": "UpdateBookEvaluation",
				"responses": {
					"200": {
						"description": "Successfully Upadated the evaluation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookEvaluationEntity"
								}
							}
						}
					},
					"400": {
						"description": "Could not update evaluation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookEvaluationError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "EVALUATION_NOT_UPDATED_ERROR",
											"message": "Evaluation with ID {evaluationId} not updated"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Could not find evaluation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookEvaluationError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "EVALUATION_NOT_FOUND_ERROR",
											"message": "Evaluation with id {evaluationId} not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Update class values available at UpdateBookEvaluationEntity from class with evaluationId\n\n\n(The volunteer must have essayModulePermission, which is checked using JWT)",
				"tags": [
					"Book evaluation"
				],
				"security": [
					{
						"jwt": [
							"essayModulePermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "evaluationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBookEvaluationEntity"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetBookEvaluationById",
				"responses": {
					"200": {
						"description": "Successfully get the evaluation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookEvaluationEntity"
								}
							}
						}
					},
					"404": {
						"description": "Could not find evaluation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookEvaluationError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "EVALUATION_NOT_FOUND_ERROR",
											"message": "Evaluation with id {evaluationId} not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get evaluation by ID\n\n(The volunteer must have bookPermission, which is checked using JWT)",
				"tags": [
					"Book evaluation"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "evaluationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteBookEvaluation",
				"responses": {
					"204": {
						"description": "Successfully delete evaluation",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"404": {
						"description": "Could not find evaluation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookEvaluationError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "EVALUATION_NOT_FOUND_ERROR",
											"message": "Evaluation with id {evaluationId} not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get evaluation by class ID\n\n(The volunteer must have bookPermission, which is checked using JWT)",
				"tags": [
					"Book evaluation"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "evaluationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/book-evaluations/by-class/{classId}": {
			"get": {
				"operationId": "GetBookEvaluationByClassId",
				"responses": {
					"200": {
						"description": "Successfully get the evaluation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookEvaluationEntity"
								}
							}
						}
					},
					"404": {
						"description": "Could not find evaluation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookEvaluationError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "EVALUATION_NOT_FOUND_ERROR",
											"message": "Evaluation with id {classId} not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get evaluation by class ID\n\n(The volunteer must have bookPermission, which is checked using JWT)",
				"tags": [
					"Book evaluation"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "classId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/book-evaluations/relevant/phrases/{date}/download": {
			"get": {
				"operationId": "GetRelevantPhrases",
				"responses": {
					"200": {
						"description": "Successfully get relevant phrases",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"format": "byte"
								}
							}
						}
					}
				},
				"description": "Endpoint to Get Relevant Phrases after a specific date download",
				"tags": [
					"Book evaluation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/notebooks/evaluation/{notebookId}": {
			"put": {
				"operationId": "SaveNotebookEvaluation",
				"responses": {
					"200": {
						"description": "Successfully Evaluated notebook",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookEntity"
								}
							}
						}
					},
					"400": {
						"description": "Notebook already evaluated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "NOTEBOOK_ALREADY_RESERVED_ERROR",
											"message": "Notebook with id {some notebook id} not found"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Notebook not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "NOTEBOOK_NOT_FOUND_ERROR",
											"message": "Notebook with id {some notebook id} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Evaluate a notebook\n\n(The volunteer must have readPermission, which is checked using JWT)",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": [
							"readPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "notebookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EvaluateNotebookEntity"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}